move single model store
