+++ src/core/modelStores/ConfigModelStore.ts
@@ -1,5 +1,5 @@
-import { SimpleModelStore } from 'src/shared/models/SimpleModelStore';
-import { SingletonModelStore } from 'src/shared/models/SingletonModelStore';
+import { SimpleModelStore } from 'src/core/models/SimpleModelStore';
+import { SingletonModelStore } from 'src/core/models/SingletonModelStore';
 import { ConfigModel } from '../models/ConfigModel';
 import { ModelName } from '../types/models';
