+++ src/core/modelStores/PropertiesModelStore.ts
@@ -1,5 +1,5 @@
-import { SimpleModelStore } from 'src/shared/models/SimpleModelStore';
-import { SingletonModelStore } from 'src/shared/models/SingletonModelStore';
+import { SimpleModelStore } from 'src/core/models/SimpleModelStore';
+import { SingletonModelStore } from 'src/core/models/SingletonModelStore';
 import { PropertiesModel } from '../models/PropertiesModel';
 import { ModelName } from '../types/models';
