--- src/shared/models/Timestamp.ts
@@ -1,5 +1,0 @@
-class Timestamp {
-  constructor(public timestamp: string) {}
-}
-
-export { Timestamp };
